requirements.txt
# To use ESP
Theano==0.9.0
# "Goldilocks" keras version. Earlier versions don't support saving models to byte streams, which we need for gRPC.
# Later models have changed the keras API in breaking ways (e.g. get_layers is removed).
Keras==2.1.6
h5py==2.9.0
numpy==1.14.2

# For gRPC
grpcio==1.16.0
grpcio-tools==1.16.0

# Jupyter notebook
notebook==5.7.4
matplotlib==3.0.2

# For XDE
pandas==0.23.4
Flask==1.0.2
gunicorn==19.9.0
click==7.0
requests==2.21.0
boto3==1.4.4

# For unit tests
nose==1.3.7
grpcio_testing==1.16.0
